[EI-112] [api] refactor: IndicatorListMapper 추가
