+++ api/src/numerical-guidance/infrastructure/adapter/indicator-list/mapper/indicator-list.mapper.ts
@@ -1,0 +1,14 @@
+import { IndicatorEntity } from '../entity/indicator.entity';
+import { IndicatorResponse } from 'src/numerical-guidance/application/query/get-indicator-list/indicator-list.dto';
+
+export class IndicatorListMapper {
+  static mapIndicatorEntityToIndicatorListDto(indicatorEntities: IndicatorEntity[]) {
+    const indicatorList: IndicatorResponse[] = indicatorEntities.map((indicator) => ({
+      id: indicator['id'],
+      name: indicator['name'],
+      ticker: indicator['ticker'],
+      type: indicator['type'],
+    }));
+    return indicatorList;
+  }
+}
